帖子模块引入redis

+++ src/main/java/com/zkm/forum/model/entity/Post.java
@@ -48,7 +48,7 @@ public class Post implements Serializable {
      * 创建用户 id
      */
     private Long userId;
-    private Integer viewCount;
+    private Double viewCount;
 
     /**
      * 创建时间
